pyproject.toml
[project]
name = "summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp>=2025.5.22",
    "python-dotenv>=1.1.1",
    "langchain>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "msal>=1.20.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]
